refactor: 프로필,마이페이지 구조 리팩토링
