+++ src/components/pages/user/profile/index.ts
@@ -1,6 +1,5 @@
 export { ProfileCard } from './profile-card';
 export { ProfileDescription } from './profile-description';
-export { ProfileDescriptionBadge } from './profile-description-badge';
 export { ProfileEditModal } from './profile-edit-modal';
 export { ProfileFollowsBadge } from './profile-follows-badge';
 export { ProfileFollowsModal } from './profile-follows-modal';
